Cache Roles method for ticket/queue objects
Roles method is called a lot of times in RT, e.g. it is called dozens of
times on ticket creation, which could cause performance issue when there
are a few custom roles involved, as the relatively slow
AppliesToObjectPredicate is called once for each custom role in each
Roles method call.

This commit caches the result of Roles at ticket/queue object level, to
improve the performance. As ticket/queue objects don't live long
generally and the cache is cleared after every Load call, we don't need
to worry about the issue that the cache might be outdated in nearly all
the cases.

We tested an RT instance with 10+ custom roles, this change could
improve the performance by about 20%.