Don't declare varaibles in modifier statements
The behavior is not defined according to perlsyn:

NOTE: The behaviour of a my, state, or our modified with a statement
modifier conditional or loop construct (for example, my $x if ... ) is
undefined. The value of the my variable may be undef, any previously
assigned value, or possibly anything else. Don't rely on it. Future
versions of perl might do something different from the version of perl
you try it out on. Here be dragons.