Try a bit harder to not wrap label tooltips
When the first row doesn't have enough space left, tooltip could be
wrapped into the second row, which looks kinda ugly when it's the only
thing that gets wrapped.

This commit sets to nowrap first and then check if the tooltip icon and
the following value column content really overlap: if they don't
overlap, then it's safe to not wrap.

Note that it doesn't fix all the cases, but makes this edge case more
rare in real world.