Fix background color for current-value dark theme
This commit fixes the current-value.form-control background color
to match the main background color for dark theme.

Because we have a rule for ".darkmode .form-control" to apply to
many different types of inputs, this more specific rule is needed to
ensure the non-input form-control elements don't also have the same
color.