Drop UpdateChecks and rename methods it calls for better clarification
These methods are used to delete passed node's related entries in
"defaults", "actions" or "rights". The "Check..." names are a bit
confusing.

You can see that the original UpdateChecks was called not just on node
deletion, but also on link deletion and node update, which is not quite
right(quite obvious under the new names). We will fix it in the
following commits.