Drop jquery.event.hover-1.0.js
It was used to delay hover action a tiny bit so menus wouldn't expand
when mouse hovers over quickly(usually not on purpose). This doesn't
work since RT 4.2 and it also breaks jQuery's hover function.

As it's not super useful anyway, here we simply drop it.

See also bf12c250dc and 87bd8a635e