Fix mobile home navigation for iOS
On iPhone, when the main navigation is collapsed into More,
clicking the Home "a" link with the sub menu expanded isn't going
back to Home. To fix it we removed the iOS specific logic which was
intentionally not clearing and setting timeout. Following the git
blame history for superfish didn't explain why the timeout reset
was being skipped for iOS.

Since it appears superfish is no longer being maintained, this
commit fixes the problem until the bug is fixed in upstream, or if
comment from the author indicates a different cause or better fix.