Create RT::Test::LDAP for common LDAP test logic
In the interests of not repeating code, let's centralise it.

This also forces the use of IPv4 for LDAP test, which solves an issue
when the net.ipv6.bindv6only sysctl is set. It turns out this was enabled
on my dev box, and is no longer a common setting.