Tweak error widget styles to make it pretty in elevator themes
The widget was not migrated for RT 5 previously and the old styles
worked badly in RT 5. Error widget generally needs 4 changes based on
new RT 5 widget:

* No outer border
* Noticeable background color
* Balanced padding
* Bold text

Note that the change to set padding in .card-body is not optional:
without this, the "Toggle visibility" feature there wouldn't be
smooth(hiding it would hide the text first, then pause a little bit,
then hide the left padding space).