Preserve quoting levels when converting HTML to plain text
When you reply to a message in RT using the rich text editor, RT creates
a text/plain version corresponding to the text/html version.  However,
quoted text is enclosed in a <blockquote></blockquote> pair and this
makes quoted text in the text/plain part difficult to discern.

This commit uses heuristics to modify the text/plain output for each
of the supported converters to detect blockquoted material and replace
it with the standard text/plain quote prefix ">".

Note that lynx does not give us any way to detect blockquoted material,
so lynx should be marked as a deprecated converter.  All of the other
supported converters do give us a way to heuristically detect blockquoted
material.