Fix styling for RT portal iframe
The RT-Portal iframe in the Admin page was being squashed and not laid out
properly due to browser style-sheets. Adding explicit style fixes the layout
to use the space on the page consistently with other components.

Added styling to allow a less squashed layout for the admin page, allowing
more room for the portal and any other content added to the right of the menu.