Add ColumnMapClassName method to RT::SharedSetting
Addresses a bug causing a fatal error in the SavedSearch dashboard component:
Can't locate object method "ColumnMapClassName" via package "RT::SavedSearch"

Commit f6cdaf8c11f252fd3e31ee85c8f8db7ad4fbbfe8 added support for multiple
types of RT::Record to the querybuilder in addition to tickets, using the
ColumnMapClassName method to provide the classname for ColumnMap to load the
correct list display. SharedSetting sub-classes aren't simple records and so
didn't have this method resulting in an error when used in dashboard
component searches.

Adding the ColumnMapClassName method to the RT::SharedSetting base class
allows objects inheriting from it to work in the same way as RT::Record based
objects, and makes any method providing a hardcoded classname un-necessary.

Fixes: I#36605