Handle encrypted outgoing emails in digest email
E.g. EmailFrequency for Alice and Bob is "Individual messages" and
"Daily digest", respectively. When RT sends an encrypted email to both
of them(they are Requestors and someone corresponds to the ticket), the
corresponding "EmailRecord" transaction is encrypted with the following
attachments:

     multipart/mixed
         application/x-pkcs7-mime
         application/x-rt-original-message

As Bob's email is deferred, the transaction also has a
"DeferredRecipients" attribute that records Bob's email address, so
rt-email-digest can include the content of the transaction.

Because the transaction doesn't have a normal textual part, the
txn->ContentObj returns nothing and previously this could cause errors
like:

    Can't call method "Content" on an undefined value at sbin/rt-email-digest line 334

This commit fixes the issue by checking if ContentObj exists before
calling Content, and also getting the textual content from the
application/x-rt-original-message part when possible.