Force update MIME encoding for ticket creation via REST 1.0
We generally don't need the encoding conversion as the fields are
already UTF-8. But SetMIMEEntityToUTF8 also massages something else,
like removing double quotes, etc. which we need here.

Note that the method is already called in other places, so it's good to
call it here too for consistency.

This is initially to fix the double-quotes issue of attachment names,
e.g.

    curl 'http://localhost/REST/1.0/ticket/new?user=root&pass=password' \
    --form 'content=Queue: General
    Requestor: root@localhost
    Subject: Some subject
    Status: new
    Attachment: "some, file.txt"
    Text: Required text field: req' --form 'attachment_1=@/tmp/test.txt'

We need to quote the filename here because it contains comma, otherwise
RT would assume it to attach 2 files, "some" and "file.txt".

Without this commit, the name saved in RT would be "some, file.txt"
including quotes.