Find full path for processing acl files on upgrade
Starting with perl 5.26, . (the current directory) is no
longer in @INC by default, which causes upgrade steps that insert
acl to fail when RT::Handle tries to require the file.
Set the full path so the acl file can be found.

This is to mirror changes in 8c61476b50, which is to fix indexes.