Search group itself for SelfService "My group's tickets"
Previously we used "WatcherGroup", which was actually to search tickets
whose roles have any user members of the group in question, instead of
the group itself. E.g. for group "security" that has "alice" as a
member.

    Ticket 1:
        Requestor: alice

    Ticket 2:
        Requestor: security

"WatcherGroup" gets both tickets(alice is also requestor of ticket 2 via
"security"), while "Watcher" only gets ticket 2, which is what we really
want.