Set path to GnuPG binary in GnuPG::Interface constructor.
Old versions of GnuPG::Interface would check the GnuPG version in the
constructor.  If the path is not supplied in the constructor, it would
attempt to execute the default gpg binary and potentially error out.

While there is a fix for the problem in GnuPG::Interface, this change
should let RT work properly with unfixed versions of GnuPG::Interface

Fixes: I#36623