Add test for when lifecycle has '-' in name
When a lifecycle name has a dash in it the regex used to determine
mappings between lifecycles will break at the "-". Meaning if we have
one lifecycle named "sales" and one named "sales-engineering" the
Mappings.html submission will interpret mappings forthe  "sales-engineering"
lifecycle as a submission for the  "sales" lifecycle.