Update mapping form submission parsing to handle '-' in lifecycle name
When a lifecycle name has a dash in it the regex used to determine
mappings between lifecycles could break at the "-". Meaning if we have
one lifecycle named "sales" and one named "sales-engineering" the
Mappings.html submission will interpret mappings for the
"sales-engineering" lifecycle as a submission for the "sales" lifecycle,
which could also result in incorrect status names.