Extract email addresses from S/MIME certificates as specified in RFC 5750.
Add a certificate with the email address only in SubjectAltName and add
a unit test to ensure that email addresses are extracted correctly.

Since RFC 5750 permits S/MIME certificates to contain no email addresses,
do not attempt to set the SMIMECertificate for an RT::User if we could
not extract an email address.