Add ability to search for disabled users
The user search form in the admin section supports a single checkbox
to indicate whether disabled users should be included in search
results. This makes it possible to include disabled users, but
impossible to exclude enabled users.

This change replaces the single checkbox with a pair of checkboxes to
independently indicate whether enabled, disabled, or both types of
users should be included. If neither checkbox is selected, rather than
always returning zero results, the include enabled checkbox is
automatically ticked for improved DWIM behavior.