Log database config overrides via PreInitLoggerMessages
If a configuration change in the database overrides one in
RT_SiteConfig, a message is logged at the warn level to note
the override. This can cause unwanted logging for command-line
scripts in cron like rt-fulltext-indexer if a user updates a config
setting via the web UI and overrides a file setting.

Add the message to PreInitLoggerMessages to log at debug instead.
The message is still available to an admin debugging a config
issue, but is suppressed for common production log levels like warn.