Add database config warnings to PreInitLoggerMessages
When RT logging for LogToSTDERR is set to warn, if a configuration
change in the database overrides one in siteconfig, a message is
output informing the admin of the override.  This is expected
behavior, but can cause unnecessary repeated emails when output
during a cronjob.  A frequent cronjob like fulltext indexer
running every couple minutes can create a lot of email.

This commit moves the database config override message to the
PreInitLoggerMessages array, which updates the output to happen at
debug instead of warn.