Add SelfServiceShowArticleSearch configuration option
This replaces a previous rights check for global ShowArticle
permission to show the article search. The new configuration
option is also global, so it has the same scope with less
overhead than a rights check.

It also makes it possible to show the search box in different
rights configurations, like when granting article access rights
at the class level.