Add a Home menu item to self service to mirror regular RT
With the new self service dashboard feature, there are now
multiple pages a user could navigate to. Provide a more
obvious way to get back to the home page. This page is the
new dashboard if SelfServiceUseDashboard is enabled or the
previous default Open Tickets page if not.

Also update the single menu item from New ticket to Create Ticket
for consistency with privileged RT.