Switch to div to not create illegal nested links
The first inner link here is the Queue name in ticket summary. Nested
links are illegal, and browsers handle them weirdly that breaks UI. E.g.
they could insert outer <a>(without content) to its descendants till the
inner <a>.

Note that "name" attribute is invalid in both <a> and <div>, here we
switch to "id" and tweak the value, which also fixes #txn-ID anchors on
the same page.

Fixes: I#37013