Fix main navigation overlap with custom logo
The right position of topactions is statically set to allow for the size
of the default RT logo. When a custom logo is uploaded not matching the
size of the default RT logo, topactions is no longer correctly
positioned.

This not only affects the layout, but the menu overflow, since
rebalanceOverflowMenu calculates based on logo width and not the
position of topactions. This can cause topactions to overlap the right
menu items.

This commit sets the position of topactions based on the logo size to
fix the layout and menu overflow if a custom logo is uploaded. It's
fired when page is fully loaded instead of logo img to make sure logo
img and #topactions are rendered.