Fix long long value overlap
Bootstrap sets the height of a value in a form-row to a static
height, which means long values are then cut off if then wrap to a
second line.

In the elevator themes, we use a combination of the classes
current-value and form-control for these values in a form-row.
This commit sets height: auto for those specific values so long
lines that overlap will not be cut off by the next line.