Ensure values for Select CFs are in the defined list
The values to which Custom Fields of type 'Select' are set should
be constrained to the list of defined values. While the Web interface
provided for this, the REST2 interface did not.

The ValidateValue method is implemented on the CustomField object
to ensure that values are valid for Select custom fields regardless
of how they are set.

Previously ValidateValue was already called in _AddCustomFieldValue,
executing the auto-created DBIx::SearchBuilder method, which
always passed by default. Since the call was already there, no
additional call to ValidateValue needs to be added to activate
this new validation.