Ensure values are valid for Select custom fields
The values to which Custom Fields of type 'Select' can be set should
be constrained to the list of defined values. While the Web interface
provided for this, the REST2 interface did not.

The ValidateValue method is implemented on the CustomField object
to ensure that values are valid for Select custom fields regardless
of how they are set.