Avoid expensive COUNT query on group dashboard rights
The call to ->Count on the groups collection to check if the
user has the right CreateGroupDashboard generates the query:

    SELECT COUNT(DISTINCT main.id) FROM Groups main
        CROSS JOIN ACL ACL_2 JOIN CachedGroupMembers...

This is triggered when dashboard rights are granted at the
group level and users do not have CreateOwnDashboard. It
is called when building the menu, so on every page load.

Since we are only checking to see if the user has the right
via at least one group, just check for one value in the
results rather than running the COUNT. This should
generate a less expensive query.