Cite explicit author for eye dropper icon
The Noun Project hosts work by many designers so they
are not always the author or owner of icons. Looking
at archived versions of pages from The Noun Project
website from when we added the icon, we found that
they are the author of the eye dropper icon we use.
Add that note to the README to clarify.

Fixes: I#37062