Provide a way to select privileged and unprivileged users in admin
It's a common confusion that the user admin page, by default,
shows only privileged users. When you search, you can get
unprivileged users in results, which can also be a bit
confusing.

Add a dropdown to allow the admin to explicitly select the
set of users they want to see. It defaults to privileged, which
was the previous behavior.

Also add the system group to search results so it is clear which
system group each user is in.