Add a rule to explicitly handle verbose|v in CLI
rt-crontool defines the verbose flag with the option to
also use the letter v like 'verbose|v'. Since that didn't
match the exception in CLI's Init, 6076e50322 included
an entry in OPT to set 'verbose' to 1 to skip the default
verbose behavior. It turned out that only worked sometimes
based on the hash order, which is random. Sometimes $verbose
would not be defined.

Add an explicit exception in the Init handler to match on the
'verbose|v' alternate to allow $verbose to be set and
used in rt-crontool.