Improve RoleGroup method to create the group if asked
We have quite a few code that calls RoleGroup just to check the members
of the group, in which case we don't want to create the group
implicitly, for efficiency and performance. Thus we add "Create"
parameter to make the creation optional.

This is to simplify role group creation for lazy ones.