Allow articles with same name in different classes
Previously, Article names had to be globally unique. This change
allows articles of the same name to exist as long as they belong to
different classes.

Note that the RT::Article::ValidateName() method is changed to allow
a class id to be passed as a second parameter, thereby altering the
behavior of the uniqueness check. If the RT::Article::Create()
method was called, it will always use the scoped uniqueness logic. If
the object was populated another way e.g. Load() then the caller is
responsible for supplying the second argument herself if that is the
desired behavior.

Fixes: I#37102