Suppress warnings with rt-fulltext-indexer --quiet
Warnings are printed to STDERR when rt-fulltext-indexer is run with
--quiet.  This can lead to frequent emails with often unnecessary
information if run through cron.

This commit updates rt-fulltext-indexer to initialize its CLI
options and RT through RT::Interface::CLI::Init rather than
directly.  This changes --quiet to set LogToSTDERR to error which
is more intuitive behavior and consistent with other RT scripts.