On Pg 9 switch key/value pair if value is CF value and key is not
For search criteria with a custom field value as the value
(right hand side), Pg 9 tries to cast *all* ObjectCustomFieldValues
content in the entire query to the type of key instead of just the
ones bound to the specified custom field, e.g.

        LastUpdated > CF.{Beta Date}

If there is an ObjectCustomFieldValue row of which content doesn't look
like a datetime (e.g. a URL text for custom field "URL"), the cast to
TIMESTAMP will fail.

This commit avoids this issue by switching key/value pair to
compare using text intead.