Sync callbacks of Users/Groups autcomplete to Principals
Principals autocomplete is a combination of Users/Groups autcomplete.
As we switched from Users to Principals for role inputs, respecting
existing callbacks helps back compatibility as customized code relying
on them won't need migration.

Here we respect not-confliting ones, including "ModifyUsersLimit",
"ModifyGroupsLimit" and "ModifySuggestion".

The "Initial" callback is not 100% compatible with the one in Users as
it doesn't have "$return" argument, and since it touches request
arguments which affects both users and groups, calling "Initial"
callback in Users here is not quite appropriate.