Let article summary take the whole width in article list
Previously we set it to 11 cols with offset 1 which didn't work as
expected because "col-offset-1" is obsolete(should be "offset-1"), so it
behaved just like "col-11".

As summary font is small and not bold, it's distinguishable enough even
without an extra offset, here we simply let it take the whole width.