Align label/value columns for Assets widget in ticket display
This is to mirror what ".current-value.form-control" looks like. The
reason we don't add "form-control" here is "ShowAssets" calls another
component "ShowRecord" to render the html, which doesn't know if it's in
a form or not(unless we pass a new argument). Adding a new css rule is
more straightforward.