Reduce space used by Current search to avoid saved search overlap
On Firefox, extra spacing around select options causes the
Current search box to push down into the Saved search section
when a Results message is shown, like when loading a saved
search. Override the default padding added on Firefox and
remove the second br to provide some additional space
to avoid the overlap.