Hide "Total" row for not-distinct results in search chart to avoid confusion
If chart is grouped by a field that contains multiple values in a
ticket(like Requestors, or multiple-value custom fields), the calculated
"Total" numbers could be misleading, e.g. "Ticket count" might not be
equal to number of tickets found. For cases like this, "Total" row
doesn't make much sense, and hiding it is more appopriate.