Reload scrubber rules when web config changes are made
RT's HTML scrubber loads its rules in the constructor and
the scrubber object is persistent via the 'state' pragma.
Some RT config options can change the rules, so if config
is changed, we need to reload.

state variables can only be modified in the scope where
they are declared, so create a flag to tell us to reload
when necessary.

Before this change, updating the ShowRemoteImages config
option would not take effect immediately in the running
RT.