Use Name as the default watcher field in search results
When we added more watcher fields in 7abb7900e4, we didn't set the
default sort field. Without this, for search result format like
"Requestors", clicking its header on search result page would cause SQL
error like:

    ORDER BY main.Requestors ASC  ... no such column: main.Requestors

Before 7abb7900e4, the default field was "EmailAddress", which is not as
good as "Name", considering by default we show "Name" for privileged
users. For unprivileged users, we do show "EmailAddress", but usually
"Name" and "EmailAddress" are the same(if not, "Name" is also shown), so
"Name" is still fine.