Update User.pm FriendlyFrom default
This PR sets a better default for FriendlyFrom as some mail filter services (like mimecast) frown on apparently-forged From headers.  This makes sense as numerous MUA clients show only the comment as the sender.  I think it makes sense here but a case could be made to perform the transformation in RT::Action::SendEmail::SetFrom instead.