Always default to no value for select type CFs on bulk update
A select type ticket custom field set as Mandatory and with a
default value set via the queue Default Values configuration
renders without the "(no value)" option. On the Bulk Update
page, it incorrectly defaults to selecting the first value in
the dropdown rather than the value defined in the queue default
configuration. This is because EditCustomFieldSelect expects any
CF defaults to be set and passed in, and Bulk Update does not
perform the lookup to check for CF default values.

If a user goes to Bulk Update to update a different field, this
results in an unexpected update of the custom field when the page
is submitted, setting it to the first value in the list for all the
tickets bulk updated.

Rather than modify Bulk Update to load the correct default value,
manually pass ShowEmptyOption to default the dropdown to "no value",
avoiding the unexpected update of the custom field completely. Users
can still select a value if they intend to bulk update that field.

The @del version of the dropdown doesn't have this issue because
it is rendered as a multiple select and as such always defaults
to having no value selected.