Retain signature if quoting content on correspondence
The UpdateContent argument to Ticket/Update.html will be quoted
as the initial value of the correspondence. It is passed, for example,
when quoteselection.js has created the ticket update link after
processing the current selection. However, the current behavior will
always hide any signature if this argument is supplied. This commit
changes the signature behavior to be the same as when UpdateContent
is not supplied.