Show user signature when quoting content from ticket display
RT-Extension-QuoteSelection noted in its "CAVEATS" that it
did not include user signatures when quoting and this behavior
was retained when it was added to core. This was likely because
it used the existing UpdateContent ARG to populate content, and
that explitily excludes signatures to avoid re-adding them when
changes are made on the update page.

Add a new arg to detect content added from quoting so we can
show signatures, making this consistent with the default
update behavior.

Also move the quoted content to a new Message ARG so it gets
positioned properly if SignatureAboveQuote is enabled.