No need to fix up attribute contents in clone mode
The PostInflateFixup is to fix subtle dependencies of attributes
contents on import, which is not necessary in clone mode.

Even worse, as different Storable versions(e.g. 2 VS 3) can produce
differnet output for the same data(because the output contains Storable
version info), if we update Content here, it could create additional
records like ContentHistory, links, etc. which will break the upcoming
imports because of id conflicts.