Allow div as a format tag for CKEditor
The format_tags configuration was originally set to simplify the
format menu in the editor, but excluding the div tag also caused
the editor to not use it when showing quoted content, breaking
the blockquote tags we add there.

The default format_tags configuration adds only a few more
format options, which doesn't really make the Format menu any less
usable, so remove the custom configuration and use the default,
which also restores div and fixes the blockquote issue.