Convert default spacing to block level p tags
In RT 5, we removed the custom config.enterMode setting of
CKEDITOR.ENTER_BR in favor of the default ENTER_P setting
since the CKEditor documentation strong discourages customizing
this option.

With this default setting, CKEditor wants content to be contained
in block level elements, so it converts the <br /> spacing around
signatures to <p>&nbsp;</p> and in doing so ends up adding an
extra line after top-quoted signatures that isn't intended.

Convert the default spacing to <p>&nbsp;</p> explicitly to
avoid the unexpected conversion in the editor. This allows us
to retain the desired spacing originally added in 9df4d3387
and do so in a way that is compatible with the CKEditor
configuration change.