Don't trigger inline edit if user clicks links, buttons or their children
This is for inline edit panels with "click" behavior. Previously only
plain links were protected. E.g. the "New Incident" link on RTIR
Incident Report page wasn't protected because the click event target is
the icon(svg) child of the link instead of the link itself.

All buttons were not protected either. Note that there is a special case
of button: selects with selectpicker are also rendered in buttons.

This commit protects both links and buttons.