Load queue object in GetDefaultQueue to make sure it's valid and visible
This is an edge case. Previously if current user didn't have "SeeQueue"
on the default queue, clicking "Create new ticket" would redirect to
ticket create page with the default queue id filled, like "#5". The
correct behavior should be to show a different queue that the user
can access, or show an error if they can't access any queues.

As we don't fill the unavailable default queue any more, now ticket
create page can fill the first available queue instead.

To be consistent with documentation, now it always returns Id if found.
Previously it could be Name if config is set to a name like "General".