Try harder to not only wrap help tooltip in labels
Labels especially for custom fields/roles could be long enough to wrap, and
it's possible that it wraps between text and the help tooltip icon, so the
last line might just contain the help icon, which looks odd.

This commit binds the text ending and help icon more tightly, to avoid the
above situation as much as possible. Note that the spaces between text span
and help icon in HTML source are intentionally removed, otherwise the wrap
could still happen right before the help icon.