Document version requirement in README.PostgreSQL
Our primary README specifies a version requirement. However, this number
is primarily based on what version PostgreSQL supported at the time of
release. The other database READMEs provide a technical justification
for their version requirements. Follow that same pattern for PostgreSQL.

Saying that we require PostgreSQL 9.1 "starting with RT 5.0.0" is
probably an oversimplification. The reality probably depends on the
versions of every piece of the stack, including libpq, DBD::Pg, and RT
itself. But RT 5.0.0 is the first release where we documented PostgreSQL
9.1+ as the version requirement (see 0e0e496580cd), so go with that.