Abstract methods to get/set/reset current interface and use them accordingly
Currently we have: API, Email, REST, Web, and CLI.

Previously we set header "X-RT-Interface" to "Mobile" for ticket
create/update requests from mobile devices. Even though "Mobile" is not
available in the new interface(as mobile devices have much larger
screens nowadays, which makes RT mobile UI not that useful as before,
not mentioning that the mobile detection is not 100% reliable), we still
set that header like before for back compatibility.