Document the true meaning of a blank DatabaseHost
The sentence "If you're running MySQL and on localhost, leave it blank
for enhanced performance." can mislead readers in a few different ways.

One, it implies that setting this blank is something only MySQL
supports. However, I've confirmed that it works the same way for
PostgreSQL. The documentation for DBD::Oracle suggests it should work
for that too, but I haven't tested it.

Two, it implies unsetting this value is *just* a performance
optimization. However, there are many ways that database server
configuration can mean that connecting over the local socket is very
different than connecting over the network:

* The socket may be unavailable to users running RT because of file
permissions, security context, or because it has been disabled in the
database server configuration.

* The socket may support different connection types than the
network. For example, in its stock configuration, PostgreSQL only
authenticates users via the `ident` method over the socket. Password
authentication is typically reserved for network connections.

You could actually use this to do a completely passwordless RT install,
with both MySQL and PostgreSQL. If you have an account with sufficient
permissions to run `rt-setup-database`, and you make DatabaseUser be
www-data/httpd, then that's sufficient to run RT. (You just need to be
careful to run cron jobs etc. also as www-data/httpd, or also grant them
their own database accounts with sufficient privileges.) This improves
security because all authentication is handled by the operating system
itself, and there's no password to leak that would let an attacker
access the database out of band. This is a little difficult to set up
today because rt-setup-database doesn't support creating the account
passwordless, but if you set it up after the fact it works.