A client terminating a connection shouldn't kill a FCGI process
When a client disconnects before processing is complete then a SIGPIPE
is sent to the FCGI process. Previously this would cause the process
to exit. Discussed on the forum here:

* https://forum.bestpractical.com/t/rt-4-4-fastcgi-processes-frequently-dying/34812
* https://forum.bestpractical.com/t/why-does-rts-fcgi-server-not-handle-sigpipe/35902

An extra note from bps: this doesn't always happen, which makes it
tricky to reproduce, but we did reproduce it and this commit fixes it.

Fixes: I#37297