Share joins for items with same op/value in watcher bundling optimization
This is mainly to reduce JOINs for performance. E.g.

    Queue = 'General' AND
    ( Requestor.Name LIKE 'root' OR AdminCc.Name LIKE 'root' OR
      CustomRole.{foo}.Name LIKE 'root' )

Previously it would create 12 JOINs(4 JOINs for each role), now it only
creates 4 in total.

Even better, since key, op and value are all the same, we can remove the
duplicates of condition clauses, thus the condition part is like:

    Users_4.Name LIKE '%root%'

instead of:

    Users_4.Name LIKE '%root%' OR
    Users_4.Name LIKE '%root%' OR
    Users_4.Name LIKE '%root%'