Suppress warning when session attachments exist
Previously toggleAttachmentWarning didn't detect attachments that were
uploaded in the current session for the current action but before the
current page load.  If the text contained a trigger word for an
attachment, and was not otherwise suppressed, this would cause the page
to display a warning to the user: "It looks like you may have forgotten
to add an attachment".  This could be triggered in several ways on
ticket creation and update.