Update the EXPORTED version
You are likely to build with this version number if you build from a Git
archive (e.g., a GitHub download), or from a Git checkout in an
environment where Git isn't available (e.g., a Docker container image
build). Keep it in sync with the current major+minor version.