Add "unknown" default priority option to priority select list
This could happen when ticket's priority is not valid any more(either
because it's an old ticket or %PriorityAsString got updated, especialy
if priority "0" is not mapped), in which case "unknown" would be shown.
Previously "unknown" was never an option of the priority select list,
which brought in 2 issues:

* Unexpected updates

On ticket basics page, if you want to update status only and not touch
priority, there was actually no way to do it.

* Set default priority value in inline edit

Inline edit submits form on input changes. Assuming "Low" is the default
value(it was selected by default), you would need to change it to some
other values(e.g. "Medium") and then change it back to "Low" to set
priority from "unknown" to "Low".