Make setting sqldump actually work
If `--sqldump` is set then the filepath will end up in `$tmp`. The later
call to `GetOptions( 'no-sqldump' => \$tmp )` will always succeed
because it is a valid getopts specification (it just doesn't set
anything). Then the `&& $tmp` condition will also pass because `$tmp`
still has a value from when ```--sqldump` earlier set it. The result is
that `$opt{'no-sqldump'}` will end up being set to the filepath of the
dump. Later on, this will seem like `--no-dump` was actually requested
and no SQL dump will take place.

Fix this by unsetting `$tmp` before it is tested again.