Support to canonicalize select values
Select values are treated case insensitively, even though previously
cases were not auto-corrected. E.g. if you set a ticket select cf value
to "foo" via email or command line(the official version is "Foo"), then
it shows as "foo" on display page, but "Foo" is actually selected on
edit pages, which is a bit confusing.

This commit corrects value cases to make it consistent with the
corresponding CustomFieldValue, to avoid the confusion.