Ignore HotList column for RT::Class on importing
HotList column has been dropped since RT 5, if we import seralized data
from RT 4, we could get the following error:

    ERROR:  column "hotlist" does not exist

This commit deletes HotList column in PreInflate to get rid of this
issue. Note that we need to do this in RT::Record instead of RT::Class
to cover both "clone" and "non-clone" modes of rt-serializer.