Add scrip to add 15 mins of time worked to ticket
This is not something that everyone will want to run automatically, and
as such, it is disabled by default. In our case, it will only get
applied to tickets in the Sales Engineering queue. If nothing else, it
will serve as an example for others to use in their own RT setups.

See T#236291 for more information.