Skip ContentHistory entries when validating Attributes
ContentHistory entries were added to the Attributes table as
a way to track deleted Saved Searches, Dashboards, and
Subscriptions. Since the attribute they were created to track
was deleted, it is correct for that original attribute to no
longer exist. Exclude these records from the referential
integrity check on Attributes.