Add RT::Article::Load
This cleanly handles being called with a non-numeric "id", using it as a
Name instead.  This prevents some database drivers from throwing errors
when they can't convert the string to a number.

Specifically, the article search box on ticket create and update can
accept a string value that doesn't autocomplete to an existing article.
This string value can be passed to Load and this previously resulted in
DB errors in the logs.