Notify user when unable to include an article
When there are enough articles that the "Include Article" dropdown is
replaced with a textbox for search (default 50, but configurable) it is
possible to submit the Ticket Create and Update pages with text that
can't be resolved to an article to be included.

Previously the user wasn't notified of this, and in the case of ticket
updates the update was committed.  Now the user will be told that the
article couldn't be loaded and the create/update will not be committed.